Modify code

+++ src/main/scala/entity/UniwareShippingPackage.scala
@@ -9,5 +9,5 @@ case class UniwareShippingPackage(
      shipping_courier: String,
      payment_method: String,
      gmv: BigDecimal,
-     quantity: java.lang.Integer
+     quantity: java.lang.Long
 )
